+++ src/runtime/components/Checkbox.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import type { CheckboxRootProps } from 'reka-ui'
+import type { CheckboxRootProps, CheckboxRootEmits } from 'reka-ui'
 import type { VNode } from 'vue'
 import type { AppConfig } from '@nuxt/schema'
 import theme from '#build/b24ui/checkbox'
@@ -8,7 +8,7 @@ import type { ComponentConfig } from '../types/tv'
 
 type Checkbox = ComponentConfig<typeof theme, AppConfig, 'checkbox'>
 
-export interface CheckboxProps extends Pick<CheckboxRootProps, 'disabled' | 'required' | 'name' | 'value' | 'id' | 'defaultValue'>, /** @vue-ignore */ Omit<ButtonHTMLAttributes, 'type' | 'disabled' | 'name'> {
+export interface CheckboxProps<T = boolean> extends Pick<CheckboxRootProps<T>, 'disabled' | 'required' | 'name' | 'value' | 'id' | 'defaultValue' | 'modelValue' | 'trueValue' | 'falseValue'>, /** @vue-ignore */ Omit<ButtonHTMLAttributes, 'type' | 'disabled' | 'name'> {
   /**
    * The element or component this component should render as.
    * @defaultValue 'div'
@@ -37,7 +37,7 @@ export interface CheckboxProps extends Pick<CheckboxRootProps, 'disabled' | 'req
   b24ui?: Checkbox['slots']
 }
 
-export type CheckboxEmits = {
+export interface CheckboxEmits<T = boolean> extends CheckboxRootEmits<T> {
   change: [event: Event]
 }
 
@@ -47,9 +47,9 @@ export interface CheckboxSlots {
 }
 </script>
 
-<script setup lang="ts">
+<script setup lang="ts" generic="T = boolean">
 import { computed, useAttrs, useId } from 'vue'
-import { Primitive, CheckboxRoot, CheckboxIndicator, Label, useForwardProps } from 'reka-ui'
+import { Primitive, CheckboxRoot, CheckboxIndicator, Label, useForwardPropsEmits } from 'reka-ui'
 import { reactivePick } from '@vueuse/core'
 import { useAppConfig } from '#imports'
 import { useComponentUI } from '../composables/useComponentUI'
@@ -60,18 +60,16 @@ import CheckIcon from '@bitrix24/b24icons-vue/main/CheckIcon'
 
 defineOptions({ inheritAttrs: false })
 
-const props = defineProps<CheckboxProps>()
+const props = defineProps<CheckboxProps<T>>()
 const slots = defineSlots<CheckboxSlots>()
-const emits = defineEmits<CheckboxEmits>()
-
-const modelValue = defineModel<boolean | 'indeterminate'>({ default: undefined })
+const emits = defineEmits<CheckboxEmits<T>>()
 
 const appConfig = useAppConfig() as Checkbox['AppConfig']
 const uiProp = useComponentUI('checkbox', props)
 
-const rootProps = useForwardProps(reactivePick(props, 'required', 'value', 'defaultValue'))
+const rootProps = useForwardPropsEmits(reactivePick(props, 'required', 'value', 'defaultValue', 'modelValue', 'trueValue', 'falseValue'), emits)
 
-const { id: _id, emitFormChange, emitFormInput, size, color, name, disabled, ariaAttrs } = useFormField<CheckboxProps>(props)
+const { id: _id, emitFormChange, emitFormInput, size, color, name, disabled, ariaAttrs } = useFormField<CheckboxProps<T>>(props)
 const id = _id.value ?? useId()
 
 const attrs = useAttrs()
@@ -106,16 +104,15 @@ function onUpdate(value: any) {
       <CheckboxRoot
         :id="id"
         v-bind="{ ...rootProps, ...forwardedAttrs, ...ariaAttrs }"
-        v-model="modelValue"
         :name="name"
         :disabled="disabled"
         data-slot="base"
         :class="b24ui.base({ class: uiProp?.base })"
         @update:model-value="onUpdate"
       >
-        <template #default="{ modelValue }">
+        <template #default="{ state }">
           <CheckboxIndicator data-slot="indicator" :class="b24ui.indicator({ class: uiProp?.indicator })">
-            <Minus20Icon v-if="modelValue === 'indeterminate'" data-slot="icon" :class="b24ui.icon({ class: uiProp?.icon })" />
+            <Minus20Icon v-if="state === 'indeterminate'" data-slot="icon" :class="b24ui.icon({ class: uiProp?.icon })" />
             <CheckIcon v-else data-slot="icon" :class="b24ui.icon({ class: uiProp?.icon })" />
           </CheckboxIndicator>
         </template>